feat: working updated weather-py example
