+++ examples/weather-py/Makefile
@@ -5,7 +5,7 @@ PROVIDER_WASM = weather_py_provider.wasm
 MCP_SERVER_WASM = mcp-http-server.wasm
 
 # Registry packages - auth is now built into transport
-TRANSPORT_PKG = fastertools:mcp-transport-http-tools@0.1.11
+TRANSPORT_PKG = fastertools:mcp-transport-http-tools@0.4.1
 
 # Python virtual environment
 VENV = venv
