+++ examples/weather-py/spin.toml
@@ -3,7 +3,7 @@ spin_manifest_version = 2
 [application]
 name = "weather-py"
 version = "0.1.0"
-authors = ["Ian McDonald <bowlofarugula@gmail.com>"]
+authors = ["bowlofarugula <bowlofarugula@gmail.com>"]
 description = "An MCP server written in Python"
 
 [[trigger.http]]
