--- examples/weather-py2/.gitignore
@@ -1,31 +1,0 @@
-# Python
-__pycache__/
-*.py[cod]
-*$py.class
-*.so
-.Python
-venv/
-ENV/
-env/
-.env
-
-# WebAssembly
-*.wasm
-
-# Generated
-wit_world/
-wit/deps/
-
-# IDE
-.vscode/
-.idea/
-*.swp
-*.swo
-*~
-
-# OS
-.DS_Store
-Thumbs.db
-
-# Temporary files
-.transport-tmp.wasm
