--- examples/weather-py2/requirements.txt
@@ -1,1 +1,0 @@
-componentize-py>=0.13.2
