--- examples/weather-py2/spin.toml
@@ -1,32 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "weather-py"
-version = "0.1.0"
-authors = ["Ian McDonald <bowlofarugula@gmail.com>"]
-description = "An MCP server written in Python"
-
-[[trigger.http]]
-route = "/mcp"
-component = "mcp-server"
-
-[[trigger.http]]
-route = "/.well-known/oauth-protected-resource"
-component = "mcp-server"
-
-[[trigger.http]]
-route = "/.well-known/oauth-authorization-server"
-component = "mcp-server"
-
-[component.mcp-server]
-# Using the composed component (provider + transport)
-source = "mcp-http-server.wasm"
-allowed_outbound_hosts = [
-    "https://geocoding-api.open-meteo.com",
-    "https://api.open-meteo.com"
-]
-key_value_stores = ["default"]
-
-[component.mcp-server.build]
-command = "make build"
-watch = ["app.py", "requirements.txt", "wit/*.wit"]
