--- examples/weather-py2/wit/world.wit
@@ -1,8 +1,0 @@
-package weather-py:provider;
-
-/// MCP tools for An MCP server written in Python
-world weather-py {
-    import wasi:http/outgoing-handler@0.2.3;
-    export fastertools:mcp/core-capabilities@0.4.0;
-    export fastertools:mcp/tools-capabilities@0.4.0;
-}
