--- examples/weather-py2/wkg.lock
@@ -1,21 +1,0 @@
-# This file is automatically generated.
-# It is not intended for manual editing.
-version = 1
-
-[[packages]]
-name = "fastertools:mcp"
-registry = "ghcr.io"
-
-[[packages.versions]]
-requirement = "=0.4.0"
-version = "0.4.0"
-digest = "sha256:b6949caeb52758254cedb2f5d9381dfcc2e918799b8cd3483cc13c5eca8521d4"
-
-[[packages]]
-name = "wasi:http"
-registry = "wasi.dev"
-
-[[packages.versions]]
-requirement = "=0.2.3"
-version = "0.2.3"
-digest = "sha256:e526c1586efc94cd148e33725139be05c4bb58ba20466d348282bd8dc3999f1d"
