+++ templates/python/content/Makefile
@@ -5,7 +5,7 @@ PROVIDER_WASM = {{project-name | snake_case}}_provider.wasm
 MCP_SERVER_WASM = mcp-http-server.wasm
 
 # Registry packages - auth is now built into transport
-TRANSPORT_PKG = fastertools:mcp-transport-http-tools@0.1.11
+TRANSPORT_PKG = fastertools:mcp-transport-http-tools@0.4.1
 
 # Python virtual environment
 VENV = venv
