+++ templates/python/content/wit/world.wit
@@ -3,6 +3,6 @@ package {{project-name | kebab_case}}:provider;
 /// MCP tools for {{project-description}}
 world {{project-name | kebab_case}} {
     import wasi:http/outgoing-handler@0.2.3;
-    export fastertools:mcp/core-capabilities@0.1.23;
-    export fastertools:mcp/tools-capabilities@0.1.23;
+    export fastertools:mcp/core-capabilities@0.4.0;
+    export fastertools:mcp/tools-capabilities@0.4.0;
 }
